Remove submitters
