+++ cg/meta/orders/api.py
@@ -13,10 +13,10 @@ from cg.apps.lims import LimsAPI
 from cg.meta.orders.ticket_handler import TicketHandler
 from cg.models.orders.order import OrderType
 from cg.services.order_validation_service.models.order import Order
-from cg.services.orders.submitters.order_submitter import OrderSubmitter
-from cg.services.orders.submitters.order_submitter_registry import OrderSubmitterRegistry
+from cg.services.order_validation_service.order_validation_service import OrderValidationService
+from cg.services.orders.store_order_services.store_order_service import StoreOrderService
+from cg.services.orders.store_order_services.storing_service_registry import StoringServiceRegistry
 from cg.store.models import User
-from cg.store.store import Store
 
 LOG = logging.getLogger(__name__)
 
@@ -27,27 +27,27 @@ class OrdersAPI:
     def __init__(
         self,
         lims: LimsAPI,
-        status: Store,
         ticket_handler: TicketHandler,
-        submitter_registry: OrderSubmitterRegistry,
+        submitter_registry: StoringServiceRegistry,
+        validation_service: OrderValidationService,
     ):
         super().__init__()
         self.lims = lims
-        self.status = status
         self.ticket_handler = ticket_handler
         self.submitter_registry = submitter_registry
+        self.validation_service = validation_service
 
     def submit(self, order_type: OrderType, raw_order: dict, user: User) -> dict:
         """Submit a batch of samples.
 
         Main entry point for the class towards interfaces that implements it.
         """
-        submit_handler: OrderSubmitter = self.submitter_registry.get_order_submitter(order_type)
-        order: Order = submit_handler.order_validation_service.parse_and_validate(
+        storing_service: StoreOrderService = self.submitter_registry.get_storing_service(order_type)
+        order: Order = self.validation_service.parse_and_validate(
             raw_order=raw_order, order_type=order_type
         )
         ticket_number: int = self.ticket_handler.create_ticket(
             order=order, user_name=user.name, user_mail=user.email, order_type=order_type
         )
         order._generated_ticket_id = ticket_number
-        return submit_handler.submit_order(order)
+        return storing_service.store_order(order)
