+++ tests/fixture_plugins/orders_fixtures/order_store_service_fixtures.py
@@ -2,19 +2,27 @@ import pytest
 
 from cg.models.orders.constants import OrderType
 from cg.services.orders.order_lims_service.order_lims_service import OrderLimsService
-from cg.services.orders.store_order_services.store_case_order import StoreCaseOrderService
-from cg.services.orders.store_order_services.store_fastq_order_service import StoreFastqOrderService
-from cg.services.orders.store_order_services.store_metagenome_order import (
+from cg.services.orders.store_order_services.instantiations.store_case_order import (
+    StoreCaseOrderService,
+)
+from cg.services.orders.store_order_services.instantiations.store_fastq_order_service import (
+    StoreFastqOrderService,
+)
+from cg.services.orders.store_order_services.instantiations.store_metagenome_order import (
     StoreMetagenomeOrderService,
 )
-from cg.services.orders.store_order_services.store_microbial_fastq_order_service import (
+from cg.services.orders.store_order_services.instantiations.store_microbial_fastq_order_service import (
     StoreMicrobialFastqOrderService,
 )
-from cg.services.orders.store_order_services.store_microbial_order import StoreMicrobialOrderService
-from cg.services.orders.store_order_services.store_pacbio_order_service import (
+from cg.services.orders.store_order_services.instantiations.store_microbial_order import (
+    StoreMicrobialOrderService,
+)
+from cg.services.orders.store_order_services.instantiations.store_pacbio_order_service import (
     StorePacBioOrderService,
 )
-from cg.services.orders.store_order_services.store_pool_order import StorePoolOrderService
+from cg.services.orders.store_order_services.instantiations.store_pool_order import (
+    StorePoolOrderService,
+)
 from cg.store.models import Application
 from cg.store.store import Store
 from tests.mocks.limsmock import MockLimsAPI
