+++ tests/meta/orders/conftest.py
@@ -1,12 +1,11 @@
-from pathlib import Path
-
 import pytest
 
 from cg.clients.freshdesk.freshdesk_client import FreshdeskClient
 from cg.meta.orders import OrdersAPI
 from cg.meta.orders.ticket_handler import TicketHandler
-from cg.services.orders.submitters.order_submitter_registry import (
-    OrderSubmitterRegistry,
+from cg.services.order_validation_service.order_validation_service import OrderValidationService
+from cg.services.orders.store_order_services.storing_service_registry import (
+    StoringServiceRegistry,
     setup_order_submitter_registry,
 )
 from cg.store.store import Store
@@ -18,18 +17,24 @@ def freshdesk_client():
     return FreshdeskClient(base_url="https://mock.freshdesk.com", api_key="mock_api_key")
 
 
+@pytest.fixture
+def order_validation_service(base_store: Store) -> OrderValidationService:
+    return OrderValidationService(base_store)
+
+
 @pytest.fixture(scope="function")
 def orders_api(
     base_store: Store,
     ticket_handler: TicketHandler,
     lims_api: MockLimsAPI,
-    order_submitter_registry: OrderSubmitterRegistry,
+    order_submitter_registry: StoringServiceRegistry,
+    order_validation_service: OrderValidationService,
 ) -> OrdersAPI:
     return OrdersAPI(
         lims=lims_api,
-        status=base_store,
         ticket_handler=ticket_handler,
         submitter_registry=order_submitter_registry,
+        validation_service=order_validation_service,
     )
 
 
@@ -39,5 +44,5 @@ def ticket_handler(store: Store, freshdesk_client: FreshdeskClient) -> TicketHan
 
 
 @pytest.fixture
-def order_submitter_registry(base_store: Store, lims_api: MockLimsAPI) -> OrderSubmitterRegistry:
+def order_submitter_registry(base_store: Store, lims_api: MockLimsAPI) -> StoringServiceRegistry:
     return setup_order_submitter_registry(lims=lims_api, status_db=base_store)
