+++ tests/services/orders/order_store_service/test_microbial_store_order_service.py
@@ -4,7 +4,9 @@ from cg.models.orders.sample_base import ControlEnum
 from cg.models.orders.samples import SarsCov2Sample
 from cg.services.order_validation_service.workflows.microsalt.models.order import MicrosaltOrder
 from cg.services.order_validation_service.workflows.mutant.models.order import MutantOrder
-from cg.services.orders.store_order_services.store_microbial_order import StoreMicrobialOrderService
+from cg.services.orders.store_order_services.instantiations.store_microbial_order import (
+    StoreMicrobialOrderService,
+)
 from cg.store.models import Case, Customer, Sample
 from cg.store.store import Store
